+++ train/src/matrix.h
@@ -14,12 +14,12 @@ namespace entity2vec {
     class matrix {
     public:
         real *data_;
-        uint32_t m_;
-        uint32_t n_;
+        uint64_t m_;
+        uint64_t n_;
 
         explicit matrix();
 
-        explicit matrix(uint32_t m, uint32_t n);
+        explicit matrix(uint64_t m, uint64_t n);
 
         void uniform(real a);
         void zero();
@@ -31,7 +31,7 @@ namespace entity2vec {
 
         void setValue(int64_t m, int64_t n, real val);
 
-        std::vector<std::pair<real, int>> findSimilarRow(uint32_t target, uint32_t k, uint32_t range_start, uint32_t range_end);
+        std::vector<std::pair<real, int>> findSimilarRow(int64_t target, uint32_t k, uint32_t range_start, uint32_t range_end);
     };
 }
